cpu/x86/pae: Fix pointer casts
Required to compile the code in x86_64, even though it's never used.

Reviewed-on: https://review.coreboot.org/c/29876
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/cpu/x86/pae/pgtbl.c
@@ -135,10 +135,10 @@ void *map_2M_page(unsigned long page)
 			memset(&pgtbl[index].pdp, 0, sizeof(pgtbl[index].pdp));
 			pd = pgtbl[index].pd;
 			pdp = pgtbl[index].pdp;
-			pdp[0].addr_lo = ((uint32_t)&pd[512*0])|1;
-			pdp[1].addr_lo = ((uint32_t)&pd[512*1])|1;
-			pdp[2].addr_lo = ((uint32_t)&pd[512*2])|1;
-			pdp[3].addr_lo = ((uint32_t)&pd[512*3])|1;
+			pdp[0].addr_lo = ((uintptr_t)&pd[512*0])|1;
+			pdp[1].addr_lo = ((uintptr_t)&pd[512*1])|1;
+			pdp[2].addr_lo = ((uintptr_t)&pd[512*2])|1;
+			pdp[3].addr_lo = ((uintptr_t)&pd[512*3])|1;
 			/* The first half of the page table is identity mapped
 			 */
 			for (i = 0; i < 1024; i++) {
